requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2
langchain-text-splitters==0.0.1
langchain-community==0.0.13
PyPDF2==3.0.1
Werkzeug==3.0.1
scikit-learn==1.4.0
numpy==1.24.0 